requirements.txt
fastapi[standard]
uvicorn
langchain
langchain-google-genai
langgraph
google-generativeai
pydantic
python-dotenv
httpx
